fix for cr 0.1
