+++ terraform_server/server.js
@@ -10,6 +10,7 @@ import tools from './tools/aiTools.js';
 import dotenv from 'dotenv';
 import { fileURLToPath } from 'url';
 import { dirname } from 'path';
+import { systemMessage } from './messages.js';
 
 dotenv.config();
 
@@ -21,40 +22,7 @@ app.use(bodyParser.text({ type: 'application/x-hcl' }));
 app.use(bodyParser.json()); // Add JSON body parser for AI policy creation
 app.use(cors());
 
-const messages = [
-	{
-		role: 'system',
-		content: `You are an intelligent assistant that converts natural language into structured Role-Based Access Control (RBAC) policies.
-
-Always reply with this JSON format:
-
-{
-  resources: [{
-    name: string,
-    actions: string[]
-  }],
-  roles: [{
-    name: string,
-    permissions: [{
-      resource: string,
-      actions: string[]
-    }]
-  }]
-}
-
-Your task:
-- Understand the user's intent, even if they don't mention "policy", "roles", or "permissions".
-- If the user describes a product, app, feature, organization, or use case — infer the implied access control.
-- Extract or invent logical roles, resources, actions, and permissions based on what the system would realistically need.
-- Use common sense and real-world examples (e.g., fintech = users, admins, money transfers, approvals).
-- Avoid asking follow-up questions unless absolutely necessary.
-- Don't use the keys "admin", "viewer", "editor" for roles.
-- In the terraform file, add field of "attributes" with empty object {} for each resource.
-- return only the table output without any other text above it.
-- Always output valid JSON and Terraform file using the permit.io terraform provider.
-- When using tools, make sure to include the formattedOutput in your response.`,
-	},
-];
+const messages = [systemMessage];
 
 app.post('/apply', (req, res) => {
 	const tempDir = `temp-${Math.random().toString(36).substring(7)}`;
